adds is active filter to brands and products

+++ src/main/java/jv/bazar/amacame/entity/Product.java
@@ -12,6 +12,7 @@ import org.hibernate.annotations.UpdateTimestamp;
 import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.annotation.LastModifiedDate;
 
+import java.math.BigDecimal;
 import java.time.Instant;
 
 @Builder
@@ -25,9 +26,9 @@ public class Product {
     private Long productId;
     private String productName;
     private Integer productStock;
-    private double productPurchasePrice;
-    private double productSalePrice;
-    private double productProfit;
+    private BigDecimal productPurchasePrice;
+    private BigDecimal productSalePrice;
+    private BigDecimal productProfit;
     private boolean isActive = true;
     @CreationTimestamp
     private Instant createdAt;
